Fix clang warnings.

+++ CesiumGltfReader/src/ImageManipulation.cpp
@@ -63,8 +63,8 @@ bool ImageManipulation::blitImage(
   size_t bytesPerSourceRow = bytesPerPixel * size_t(source.width);
   size_t bytesPerTargetRow = bytesPerPixel * size_t(target.width);
 
-  size_t requiredTargetSize = targetPixels.height * bytesPerTargetRow;
-  size_t requiredSourceSize = sourcePixels.height * bytesPerSourceRow;
+  size_t requiredTargetSize = size_t(targetPixels.height) * bytesPerTargetRow;
+  size_t requiredSourceSize = size_t(sourcePixels.height) * bytesPerSourceRow;
   if (target.pixelData.size() < requiredTargetSize ||
       source.pixelData.size() < requiredSourceSize) {
     return false;
